Complete fileCheck & IPValidation
*NOTE THIS IS ALL PRE-TESTING*
This all *should* work but I haven't tested it yet. 
Used regex to specifically print name and modify date for the compromised files from the output of the command. Also created and returned a list containing the paths of the compromised files relative to the home directory. 

Also used regex to validate for IP address to ensure it was the correct format.